feat: 4. 삭제 예외처리 - 400 RemainArticlesException

+++ src/main/java/com/example/bcsd/service/MemberService.java
@@ -1,10 +1,11 @@
 // 모든 비즈니스 로직들을 작성한다.
 package com.example.bcsd.service;
 import com.example.bcsd.domain.Member;
-import com.example.bcsd.dto.ArticleDTO;
 import com.example.bcsd.dto.MemberDTO;
 import com.example.bcsd.exception.DuplicateEmailException;
 import com.example.bcsd.exception.MissingFieldException;
+import com.example.bcsd.exception.RemainArticlesException;
+import com.example.bcsd.repository.MemoryArticleRepository;
 import com.example.bcsd.repository.MemoryMemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,10 +16,12 @@ import java.util.List;
 @Service
 public class MemberService {
     private final MemoryMemberRepository memberRepository;
+    private final MemoryArticleRepository articleRepository;
 
     @Autowired // 생성자 호출할 때 MemberRepository(실제로는 구현부인 MemoryMemberRepository)를 넣어줌!
-    public MemberService(MemoryMemberRepository memberRepository) {
+    public MemberService(MemoryMemberRepository memberRepository,  MemoryArticleRepository articleRepository) {
         this.memberRepository = memberRepository;
+        this.articleRepository = articleRepository;
     }
 
     public Member create(MemberDTO req) {
@@ -56,8 +59,12 @@ public class MemberService {
     }
 
     public boolean delete(Long id) {
+        if (articleRepository.existsByAuthorId(id)){
+            throw new RemainArticlesException("사용자가 작성한 게시글이 남아있어 삭제가 불가능합니다. member_id =" + id);
+        }
         return memberRepository.deleteById(id);
     }
+
     private void validateForCreate(MemberDTO req) {
 
         if (req.getName() == null || req.getName().isBlank()) {
@@ -71,5 +78,8 @@ public class MemberService {
         if (req.getPassword() == null || req.getPassword().isBlank()) {
             throw new MissingFieldException("password은 null일 수 없습니다.");
         }
+    }
+    private void validateForDelete(MemberDTO req) {
+
     }
 }
